The DungeonMaster refrence is now sent to an asset's constructor.
Previosly it was being passed around everywhere.  I had attempted
to just make DungeonMaster a singleton, and have assets that needed
the DM to just grab the singleton instance.  That ended up creating
a circular refrence because the DM already requires the assets,
a singleton would mean the assets would also require DM.